transition-view: Directly store the RigNode pointer for selected nodes
Previously the list of selected nodes in the transition view was only
storing the time offset for each node. Every time the node was
manipulated it would search in the path for the node with the given
time. Now that RigPath has a function to directly search for nodes
with a documented lifespan for the node, we can just directly store a
pointer to node. This is slightly easier to manage and avoids a few
list iterations.

The undo journal now also takes a pointer to a node to remove instead
of a time value. The details of the node (ie, the time and the value)
are still copied out of the node when logging so that the node can
safely be destroyed.