transition-view: Allow dragging path nodes
The button down handler for the transition view is now delayed so that
it can do one of three operations depending on what the next event is.
If the next event is a button up it will just select the node under
the pointer as before. If the mouse is moved it will either start
moving the timeline progress or if there is a node under the pointer
it will start dragging all of the selected nodes.

The drag is restricted to a limited range so that moving the nodes
won't make any of them go off the end of the timeline or cross a
neighbouring node.

The undo journal now has an extra operation to track moving a bunch of
nodes. The transition view uses this to commit the move when the mouse
button is released.