rut-prop-inspector: Fixing storing the closure for the animated cb
The prop inspector struct has a field to store the closure for the
global animated callback on the property context. However it wasn't
updating this field when the callback is added so it would never
actually get detached. This was causing it to crash when the selected
entity is changed and the animated state of a property is changed.