Tidy up the flags on RutPropertySpec
Previously the flags for RutPropertySpec were READABLE or READWRITE.
It seems a bit confusing to have two ways to mark a property as
readable and that you can't make a write-only property. This patch
changes it so that there are READABLE and WRITABLE flags and READWRITE
is just the combination of the two. It also updates the flags on all
of the property specs.

There is now a warning in rut_property_init to verify that the
property spec is either readable, writable or both and that it has a
getter or a setter depending on the flags.

Some of the properties in RutText were neither readable nor writable
which is a bit useless so this adds a getter for them and makes them
readable.