Move responsibility for toggling animated state out of inspector
The property inspector no longer directly updates the animated state
of a property. Instead it is left up to Rig to do it in response to
the animated state callback. Instead of reading the property's
animated state, the inspector now just expects to be told about its
state via the new rig_inspector_set_property_animated. Rig now
iterates all of the properties and tells the inspector about the
animated ones whenever it creates a new inspector.

The idea is that the animated state should be tracked by the
transition rather than being part of the RigProperty itself because
eventually there will be multiple transitions. The property inspector
can not directly know about the transition's state because it is part
of Rig not Rut.