transition: Pass the property data directly in the foreach callback
Instead of passing the property, the path and the constant value as
separate arguments to the callback for transition_for_each_property it
now just passes a pointer to the RigTransitionPropData struct. This
makes it easier if we later want to add more properties to the struct
such as whether the property is animated.