Manage properties' animated state in the transition not RutProperty
The ‘animated’ flag for a RutProperty has been removed and instead it
is stored as part of the RigTransitionPropData. Instead of a global
callback in the property context to report whenever any property's
animated state changes, there is now a callback to report operations
on the transition. Currently the only operations are that a property
has been added or that its animated state has changed.

This should make it easier to put the property data for deleted
entities into the journal. Otherwise it would be difficult for the
transition view to recognise that the animated properties in the
journal aren't part of the transition and so shouldn't be displayed.
It should also make easier to eventually support multiple transitions.