material: be smarter about constructing from an asset
This changes rut_material_new to not assume that the given asset is a
texture asset and also cope with a NULL asset pointer. If an asset is
specified then it checks if it's a texture, alpha mask or normal map
so it can be correctly associated with the material.