Expose depth of field properties via camera
This add focal_distance and depth_of_field properties to RutCamera which
are now set as uniforms on the pipelines used for the depth of field
effect instead of the previous 30.0 and 3.0 constants respectively.

Since the current interface doesn't provide a way to select the camera
entity this adds a hacky key-binding for ctrl-c to select the camera.