undo-journal: Preserve the entity's transition data on delete
When an entity is deleted, all of the per-property data stored in the
transition is now also captured in the undo entry. This is reloaded in
the transition when the operation is undone.

Instead of having a single operation for delete with a flag to mark
whether it is inverted, there is now a separate operation for adding
which shares the same data struct. Inverting either operation just
involves tweaking the operation enum.

Rig will now additionally log entity additions in the undo journal
using the new add operation.