Add rig_set_selected_entity
Instead of directly manipulating the data->selected_entity pointer,
there is now a global function to update it. This will cause a redraw
and update the inspector. This is now called when an entity is deleted
so that it won't be left showing the rotation tool on a non-existant
actor. The undo journal also now uses to reselect delete actors when
the deletion is undone.