Ignore drags that don't actually move the entity
Previously if the user clicks on an entity and then lets go without
moving the mouse then it would add a path node at this position.
Presumably in this case the user just wants to select the node so the
path node is unnecessary and annoying. This patch just makes it so
that the entity translation grab keeps track of whether it has
received a move event. This is passed to the grab complete callback
which won't bother committing the position change if nothing moved.