Bail out if no filename argument is given
If no filename argument is given then Rig would previously crash
because nothing would set an assets directory. The rest of the code
assumes there is always an assets directory. Using Rig without a
filename is currently not useful because there would be no way to save
the created transition so we might as well just bail out in that case.

This patch also changes it to use a global variable to specify the
filename instead of always reading from _rig_handset_remaining_args.
This should make it slightly easier to maintain if we ever want to
change were it gets the filename from.