Create an initial node when a property is first animated
When a property is marked as animated and its path is empty (ie, this
is probably the first time it has been animated) then Rig will now
create an initial path node at the current transition progress with
the property's current value. The new path node is logged within the
same entry in the undo journal so that a single undo operation will
both remove the initial node and unmark the property as animated.