Fix undoing a constant property change
When a constant value is modified on the transition and then later
undone, it was only setting the object's current value and not
actually modifying the transition. This fixes it to update the
constant value and then ask the transition to update the object's
value based on the new transition value.