Fix a warning about using = instead of ==
The code to handle input for an asset has a variable to report whether
the event was handled or not. Part of the code was trying to check
this value to decide whether to queue a redraw or not. However, it was
assiging the value instead of checking it which would mean it would
always swallow all the events and always queue a redraw.

I'm not exactly sure if there are any bugs caused by always swallowing
the events but it seems worth cleaning up anyway.