Pregenerate the snippets used for entity pipelines
When creating a pipeline for an entity, it would previously always
allocate some new snippets and attach them to the pipeline. The
program cache in Cogl never looks at the contents of a snippet but
instead only compares the snippet pointers. This meant that each
entity would end up with its own program, even if it is using exactly
the same shader as another entity. This would quickly end up
generating a warning from Cogl complaining that there are more than 50
programs.

To fix this, the renderer code now generates all of the snippets that
it is going to use upfront and attaches them to the RigData struct.
The exact same snippets are attached to all pipelines so they can
easily be reused.

The renderer also uses some other snippets that I haven't touched
because they are already attached to a template pipeline and only
created once.