Add an HSV color picker widget
This adds a color picker widget which is used by the property
inspector to configure color properties.

The widget is split into two parts. The widget shown by the property
inspector is a simple button which shows the currently selected color.
When the button is pressed a second widget is shown which can select
the color. The second widget works like a modal window. The button
widget takes a grab to detect clicks outside of the modal window so it
can destroy it once the user interacts with something else.

The picker widget shows a circle to select the hue and saturation
components and a vertical gradient to select the value component.
These two shapes are drawn using a texture with all the calculated
values in. The values are calculated on the CPU in a completely
non-optimised way but it seems to perform well enough. The textures
are recalculated as the components change.