Use a union to add some type safety to the property getters&setters
The getter and setter pointers in the property spec are now stored in
an union with a member for each of the possible types of functions so
that the compiler can give a warning if the wrong type is used in the
initialiser. For example, instead of initialising the function with:

  {
    /* ... */
    .type = RUT_PROPERTY_TYPE_FLOAT,
    .setter = rut_explosion_get_megatonnes,
    /* ... */
  }

It would now be:

  {
    /* ... */
    .type = RUT_PROPERTY_TYPE_FLOAT,
    .setter.float_type = rut_explosion_get_megatonnes,
    /* ... */
  }

There is also an 'any_type' which is a void* pointer as before. This
is still used for enum types and pointer types because it would
probably not be a good idea to force those functions to return int or
void* in order fit in the function pointer.

This patch also fixes a fair few getters for color types which were
returning the color in an out-param instead of returning a constant
color pointer. This was causing a crash when the camera was selected
and the inspector would try to read one of the color values.