load-save: Update properties from the transition state on load
After loading all of the data, it will now set all of the properties
on the object that the transition knows about, either using the
transition's path or the constant value depending on whether the
property is animated. Previously the properties would just be left at
whatever values they had when the document was saved. These might not
be consistent because after loading the timeline progress will be
reset back to zero.

Eventually we might want to get rid of the hand-coded attributes for
components and entities and save particular values. Anything that can
be modified will end up having a property in the transition anyway so
the property attributes are redundant. Saving the values by
introspection will result in having to write a lot less manual code.

Ideally in the long term we don't want to store the constant values in
the transition because this makes it difficult to have multiple
transitions. Whatever new mechanism we choose should still save the
properties by introspection though.