load-save: Allow saving property data for components
All of the components are now assigned an id when loading and saving.
This allows the property data in the transition to be associated with
the correct object. Previously the property data would be saved with
an id of "0" and then it would fail trying to load the data.

The data format of the saved files has now changed because the
components now need an id attribute. However, to maintain
compatibility the loader will accept a component with a missing id but
there is a comment to say that we should remove this when we get to a
point where we can break the file format.