transition-view: Add support for selecting nodes with a bounding box
When the user drags the cursor over the transition view with the shift
key held down it will now draw a rectangle for that region. When the
mouse button is released it will select all of the nodes within that
box. Any nodes outside of the box that were already selected remain
selected.