transition-view: Group properties by entity not component
When a property for a component is added it will now add the property
to the group for the component's entity instead of making a separate
unnamed group for the component. The properties are sorted by name
within a group and the groups are sorted so that the entity properties
are always first. Other components are just sorted by the their
address because it probably doesn't really matter what the order is as
long as it is consistent.