transition-view: Use explicit variables to track the area for nodes
Previously there was a variable to store the width of the area for the
property labels. Everything that wanted to know the position of the
area for nodes would calculate it by subtracting the width of the
labels from the total width. Instead of that there are now separate
variables for the x position and the width of the area for the nodes
so that later it will be easier to move the area.