ui-viewport: Don't expand the document underneath the scroll bars
Whether or not the scroll bar appears affects the total size of the
viewport. This should also affect the size of the document when the
expand properties are enabled. To make this easier to implement, the
length and viewport of the scroll bars are now configured during the
pre-paint callback instead of immediately when the document size
changes. If there is a sync widget then the scroll bars are configured
using the preferred size of the widget instead of the size of the
document. That way the actual size used for the document can depend on
the presence of the scroll bars.

The doc x and y properties have been changed so that they no longer
depend on the scroll bar position. Previously this was creating a
circular dependency. Although that would work fine, it seems simpler
to make the doc position properties simply be a proxy to retrieve the
scroll bar position. That way there is only one canonical location for
the position and we only need one bit of code to handle the clamping.
Pressing the page up and down keys now effectively just updates the
scroll bar position. This in turn causes it to scroll.