scroll-bar: Reclamp the offset when the length or viewport changes
When the length or viewport of the scroll bar changes it will now
reclamp the offset position. This fixes problems were the offset would
get into an inconsistent state and a viewport could be shown scrolled
lower than its maximum.

Additionally, the setter for the offset now checks whether the same
value is being set after it is clamped to make it more robust.