Add the inspectors to a RutBoxLayout
The widgets for the inspectors are now added to a box layout so that
they will automatically all be allocated the same horizontal width.

The list of component inspectors has been renamed to 'all_inspectors'
and now also includes the main entity inspector. This simplifies some
of the code to handle the inspectors.