Use rut_paint_graph_with_layers to paint the UI graph
rut_paint_graph_with_layers is used to allow widgets to paint in
multiple layers so that certain parts of the widget are over
everything else. The RigDropDown was trying to use this but as nothing
was setting the layer number it wouldn't actually paint anything. This
was causing an odd gap in the inspector for the light.