undo-journal: Suppport an entry representing a collection of entries
There is now an undo operation that represents a collection of entries
that will be undone as if they were a single operation. The collection
is actually stored as another RigUndoJournal so that the normal
functions can be used to add to the journal. The inverse of the entry
just creates another journal with all of the inversions of the
subentries in reverse order.

This could be used for example when toggling an animated button.
Sometimes this needs to also add the initial path node when the path
is empty. It could be more convenient to use the normal path adding
operation to avoid duplicating code to handle it.