Use a subjournal to log toggling the animated state
When the animated state of a property is initially enabled it will add
a node in the path of the current transition with the current value.
Previously the undo entry for toggling the animated state was aware of
this and tracked the node's value. To reduce code duplication, this
now uses a subjournal entry with the two actions instead. They are
still undone as one single action.

This reverts commit 9a4af9d899cab8bbdb08bd79130c4d02a59db905.