Remove the pipeline layer user matrix
The pipeline layer user matrix is not used very often and the way it
is currently implemented comes at a non-trivial cost. Any pipeline
layer that requires the big state needs to have a CoglMatrix added to
its state. The generated vertex shader will always multiply the
coordinates by the matrix even if it is just the identity matrix.

The same functionality can be achieved on programmable hardware using
a shader snippet. The snippet hook for the texture coordinate
transform is still available to make this convenient.

If we later decide that this is a really useful feature for fixed
function hardware then we can add it back in and try to make it work
more efficiently. In the meantime however it seems like a good idea to
get rid of the API now before Cogl 2.0 is released and the API becomes
fixed.

Reviewed-by: Robert Bragg <robert@linux.intel.com>