Define rut_cogl_context in a C source not just a header
rut_cogl_context was only being defined in rut-global.h which meant
that every source file that included it would get its own global
symbol. The compiler on OS X was getting upset about this. To fix it
the header now just declares it as external and the symbol is now
defined only once in rut.c