Make a separate event for text strings
SDL2 has a separate event that the keyboard generates when keypresses
mean that the user has typed some text instead of just stuffing a
single unicode character into key down events. I think this is a
better model because it means there doesn't have to be a 1:1 mapping
between keypresses and characters.

This patch makes Rut adopt the same model. There is now a new event
type called RUT_INPUT_EVENT_TYPE_TEXT. There is only a single property
for these events retrieved with rut_text_event_get_text(). This
returns a null-terminated UTF-8 string containing the text.

On SDL2, this function just directly returns the text contained in the
text input event. On SDL1 the unicode member of the keydown event is
converted to UTF-8 and blatted over the top of the data for the
keysym.

The RutText widget has been changed to handle both types of events.
Previously on SDL2 the unicode property of the event wasn't filled in
so text widgets didn't work at all.