renderer: Use a fuzziness threshold when discarding opaque pixels
Rendering 3D models such as the big arrow representing the light on
the Mac Mini was causing sparkly artefacts. This appears to be
happening because the interpolation of the alpha value across the
primitive does not always result in a value that is exactly 1.0 even
if all of the vertices in the primitive are 1.0. In that case some of
the pixels of the model would end up being rendered with the blended
pipeline instead of the unblended one. The blended pipeline does not
use depth testing so sometimes parts of the back of the model would
show through.

This patch fixes it by comparing to >= 0.9999 instead of == 1.0. As
mentioned in a comment in the patch however, it might make more sense
to just disallow transparency in a 3D model because you would need to
sort all of the individual triangles in the model according to depth
to get it to work anyway.