l10n: Remove the internal copy of glib from the list of gettext files
We probably don't want to include the internal copy of glib in the
list of files to scan for translatable strings because these strings
already exist in glib itself and it is a waste of translator's time to
translate them again.

If we wanted to make this work properly, maybe we could adopt a system
where we also copy the translated strings from glib into cogl and then
merge them into the message catalog during the build. However it
doesn't seem like translated strings would be that useful in
situations where --disable-glib is used, so maybe it's not worth
bothering with.

https://bugzilla.gnome.org/show_bug.cgi?id=685238

Reviewed-by: Robert Bragg <robert@linux.intel.com>