Add a resize handle on Mac OS X
In an NSWindow on OS X, there is a special property to make it
automatically draw a little resize handle in the bottom right corner
of the window. This patch adds the necessary call to make it set that
property. The call has to be written in Objective-C so there is now a
separate function in a .m file to initialise on onscreen framebuffer
for OS X. There are some tweaks to the configure script and the
Makefile to allow building ObjC files. Notably, some of the CFLAGS are
now moved to CPPFLAGS so that they will be passed to the ObjC compiler
too.

When using GL with an NSWindow, the framebuffer blats over the resize
handle drawn by Cocoa. It is however still functional even though it
is invisible. To workaround this, Rig now manually draws the resize
handle using a RutImage.