Add a file menu on OS X for opening a different file
The OS X code now adds a file menu to save the current document and to
open a different one. The currente filename has been moved from a
static global variable to being a member of RigData. Rig doesn't
really cope with changing the file yet so this actually just makes it
crash.