Show alert dialogs from the automatic update on OS X
When Rig is started on OS X it will now check for specific files that
the automatic updater script leaves to report status. One of the files
is used to report errors and the other is used to report a successful
update. Both of these just cause it to display a modal dialog using
Cocoa.