osx: Open new files by restarting the main loop
When the open file dialog is used on OS X, it will now store the
filename in the RigData struct and then quit the main loop. The main()
function will now detect this filename and restart. This is a bit of a
sledgehammer to avoid having to work out how to clean up the graph
properly when a new file is loaded.