camera: Adds zoom property
This adds a zoom property to the RutCamera component that affects the
projection matrix by changing with the width and height of the frustum
according to a camera->zoom factor. This is now used by the editor
camera to handle zoom navigation so when we are in play mode zooming no
longer affects the depth of field effect.